requirements.txt
PyQt5>=5.15.0
httpx>=0.25.0
python-dotenv>=1.0.0
